wind down on generics
